Criado o orderItemService e orderItemController. Melhorias no OrderService

+++ order/src/main/java/br/com/lucio/order/application/service/OrderService.java
@@ -1,17 +1,14 @@
 package br.com.lucio.order.application.service;
 
-import br.com.lucio.order.application.exception.ServiceException;
-import com.fasterxml.jackson.databind.deser.std.UUIDDeserializer;
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import br.com.lucio.order.application.dto.OrderDTO;
+import br.com.lucio.order.application.exception.ResourceNotFoundException;
 import br.com.lucio.order.domain.entity.Order;
 import br.com.lucio.order.domain.entity.Status;
 import br.com.lucio.order.domain.repository.OrderRepository;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
@@ -28,24 +25,29 @@ public class OrderService {
         Order order = findOrder(id);
         return modelMapper.map(order, OrderDTO.class);
     }
+
+    private Order findOrder(UUID id) {
+        return orderRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException(String.format("Order not found with id %s", id)));
+    }
+
     @Transactional
     public OrderDTO createOrder(OrderDTO orderDTO) {
         Order order = modelMapper.map(orderDTO, Order.class);
         order.setStatus(Status.REGISTERED);
-        Order orderSet = order;
-        order.getItems().forEach(item -> item.setOrder(orderSet));
-        order = orderRepository.saveAndFlush(order);
+        order.getItems().forEach(item -> item.setOrder(order));
+        orderRepository.saveAndFlush(order);
         return modelMapper.map(order, OrderDTO.class);
     }
 
     @Transactional
     public void deleteOrder(UUID id){
-        findOrder(id);
+        if(!orderRepository.existsById(id)) {
+            throw new ResourceNotFoundException(String.format("Order not found with id %s", id));
+        }
         orderRepository.deleteById(id);
+        orderRepository.flush();
     }
-    private Order findOrder(UUID id) {
-        return orderRepository.findById(id).orElseThrow(() ->
-                new EmptyResultDataAccessException(String.format("Order not found with id %s", id), 1));
-    }
+
 
 }
